Remove unittest::suite
The unittest::suite class has been removed simply because it is
unnecessary. MNMLSTC Unittest relies on CMake and CTest for how it is
built and used. Any 'suites' are the tests that are placed directly
within the generated executable, while groups of suites are handled at
the CTest add_test level. For this reason, the suite type has been
removed.