Modify the unittest::v1::exception interface to provide a 'type' string
This will allow for the test runner to better report what kind of
exception was caught, and display better information. e.g.::

    assert_is: lhs is not rhs value in statement #N

Modify unittest::v1::skipping to be non-inheritable, and a base.
Modify unittest::v1::identity_crisis to be non-inheritable and a base.

Due to both of these only ever being used to print a message (that is,
the eventual test runner will handle the final formatting of messages),
they have no reason to hold their type and statement number.