Remove assert_is_instance and related tests
Due to a few ambiguities that pop-up when the compiler is able to
determine at compile time whether a given type *is* an instance of
something, I've removed the is_instance check. If a user has RTTI
enabled, they can do a check with typeid, type_index, or dynamic_cast
with a catch around it. (However, the dynamic cast is the cause for this
issue). I *could* use an #ifdef UNITTEST_ENABLE_RTTI check within the
identity class, but this breaks the 'clean' interface I'm shooting for
(even though at the moment it is hackneyed and all over the place)