Modify unittest::repr function to allow compilation on gcc.
For some reason, the code just wasn't working right. In hindsight, this
implementation is much cleaner, but I feel slightly unnerved, and feel
that I've gotten lucky as to why it is compiling just fine. If this
stuff breaks in the future (which it might because gcc seems to break
all the time!), I'm going to have to look into a long term solution for
the *very* basic printing of unknown types. (Possibly overriding
std::to_string with the <object at 0x> syntax?)